Move download script to inside package
