+++ docs/Home.py
@@ -1,5 +1,5 @@
 import streamlit as st
-from download_prerequisites import download_data
+from policyengine_us_data.utils.docs_prerequisites_download import download_data
 
 download_data()
